Split create_h5parm() into losoto.import "library"
